+++ mojo/public/java/system/src/org/chromium/mojo/system/Handle.java
@@ -56,5 +56,6 @@ public interface Handle extends Closeable {
      * Releases the native handle backed by this {@link Handle}. The caller owns the handle and must
      * close it.
      */
-    public long releaseNativeHandle();
+    public int releaseNativeHandle();
+
 }
