+++ services/shape_detection/android/java/src/org/chromium/shape_detection/InterfaceRegistrar.java
@@ -14,12 +14,12 @@ import org.chromium.shape_detection.mojom.TextDetection;
 
 @JNINamespace("shape_detection")
 class InterfaceRegistrar {
-    static MessagePipeHandle messagePipeHandleFromNative(long nativeHandle) {
+    static MessagePipeHandle messagePipeHandleFromNative(int nativeHandle) {
         return CoreImpl.getInstance().acquireNativeHandle(nativeHandle).toMessagePipeHandle();
     }
 
     @CalledByNative
-    static void bindBarcodeDetectionProvider(long nativeHandle) {
+    static void bindBarcodeDetectionProvider(int nativeHandle) {
         // Immediately wrap |nativeHandle| as it cannot be allowed to leak.
         MessagePipeHandle handle = messagePipeHandleFromNative(nativeHandle);
 
@@ -33,13 +33,13 @@ class InterfaceRegistrar {
     }
 
     @CalledByNative
-    static void bindFaceDetectionProvider(long nativeHandle) {
+    static void bindFaceDetectionProvider(int nativeHandle) {
         FaceDetectionProvider.MANAGER.bind(
                 new FaceDetectionProviderImpl(), messagePipeHandleFromNative(nativeHandle));
     }
 
     @CalledByNative
-    static void bindTextDetection(long nativeHandle) {
+    static void bindTextDetection(int nativeHandle) {
         // Immediately wrap |nativeHandle| as it cannot be allowed to leak.
         MessagePipeHandle handle = messagePipeHandleFromNative(nativeHandle);
